Implemented CGI Web Server
Incomplete feature of CGI addition--script does not run.
