--- JavaScript/googleCSE.js
@@ -1,9 +1,0 @@
-(function() {
-    var cx = '011139275606927959748:wlk5jjddrea';
-    var gcse = document.createElement('script');
-    gcse.type = 'text/javascript';
-    gcse.async = true;
-    gcse.src = 'https://cse.google.com/cse.js?cx=' + cx;
-    var s = document.getElementsByTagName('script')[0];
-    s.parentNode.insertBefore(gcse, s);
-})();
