+++ cgi-bin/NewEggs.py
@@ -1,35 +1,44 @@
-'''
-This file is mean to use a test file to avoid damaging main code file.
-Removing this file project has no effect to the project
-
-'''
-
-'''
+#!/usr/bin/env python3
 import selenium.webdriver as webdriver
 from bs4 import BeautifulSoup
 import csv
+import cgi
 
 def get_link(search_term):
     # Make sure to move chrome driver to project folder
     chromeDriver = '/usr/local/bin/chromedriver'
+    print('Moving Chrome Driver...')
+
+    # Add options to chrome
+    options = webdriver.ChromeOptions()
+    options.add_argument('headless')
+    options.add_argument('window-size=1920x1080')
+    print('Adding options to Chrome...')
 
     # Website Links
     url = "https://www.bestbuy.com/"
     newegg_url = "https://www.newegg.com/?nm_mc=KNC-Askcom&cm_mmc=KNC-Askcom-_-Branding-_-NA-_-NA"
 
     # Selenium Setting chrome driver and access website
-    browser = webdriver.Chrome(chromeDriver)
+    browser = webdriver.Chrome(executable_path = chromeDriver, chrome_options = options)
+
+    # Add options
+    options = webdriver.ChromeOptions()
+
+    options.add_argument('headless')
+
+    # Get the url
     browser.get(newegg_url)
+    print('Fetching the URL...')
 
     # Find search bar input content
     search_bar = browser.find_element_by_id("haQuickSearchBox")
+    print('Targeting search bar...')
 
     # Pass in Search term and do a automatic search
     search_bar.send_keys(search_term)
     search_bar.submit()
-
-    # Creating a csv file and put settinf as w ( write )m should use a for setting
-    #f = open("Products.csv", "w")
+    print('Querying your search item...')
 
     # Grabbing all HTML Code
     html_source = browser.page_source
@@ -37,16 +46,13 @@ def get_link(search_term):
     # Setting pass in text code as html code
     soup = BeautifulSoup(html_source, "html.parser")
 
-    #Create header for csv file
-    headers = "product_name, price\n"
-    #f.write("")
-
     # List to store data
     products_info = []
 
     # Soup object searching through products containers
     container_box = soup.findAll("div", {"class" : "item-container"})
 
+    # For loop to find name, price, and link through the container section
     for contain in container_box:
         # Name section
         name_box = contain.find("a", {"class":"item-title"})
@@ -61,48 +67,23 @@ def get_link(search_term):
         NewUniTempPrice = UniTempPrice.replace('\xc2\xa0\n', ' ').strip('\n\xe2\x80\x93').strip(' ')
         FinalUniTempPrice = NewUniTempPrice.replace('\xc2\xa0',' ').strip(' ')
 
+        # Add to product list
         products_info.append([UniTempName, FinalUniTempPrice, UniTempLink])
+    print('Fetching data...')
 
-
+    # Create, open csv file
     with open('products.csv', 'wb') as myfile:
-        wr = csv.writer(myfile, quoting = csv.QUOTE_ALL)
+        # wr = csv.writer(myfile, quoting = csv.QUOTE_ALL)
+        # Write the header
         wr.writerow(("Name", "Price", "URL"))
+        # List through the tuple and write to file
         for info in products_info:
             if(search_term.upper() in info[0].upper()):
                 wr.writerow(info)
+    # close the csv file
     myfile.close()
-    '''
+    print('Outputing to CSV file...')
 
-    '''
-    with open('products.csv', 'wb') as myfile:
-        wr = csv.writer(myfile, quoting = csv.QUOTE_ALL)
-        wr.writerow(("Name", "Price", "URL"))
-        for key in productDictionary.keys():
-            neweggs_name = key
-            inputCSV_row = neweggs_name
-            for value in productDictionary[key]:
-                newsegg_price_url = value
-                inputCSV_row = inputCSV_row + "," + value
-            inputCSV_row = inputCSV_row + "\n"
-            wr.writerow(inputCSV_row)
-    myfile.close()
-
-    '''
-    # Creating a dictionary and bound the product name and price together
-
-    '''
-    print ("Name list size: ",len(name))
-    print ("Price list size: ", len(price))
-    print ("Link list size: ", len(link))
-    print("--------------------")
-    print(name)
-    print ("--------------------")
-    print price
-    print("----------------------")
-    print(link)
-    print("----------------------")
-    productDictionary = dict(zip(name,zip(price, link)))
-    print productDictionary
-    '''
-
-get_link("alienware")
+def __init__(self):
+    HtmlFormSearch = cgi.FieldStorage()
+    get_link(HtmlFormSearch.getvalue('search'))
