+++ cgi-bin/hello.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+import cgi
+print 'Content-Type: text/html\n\n'
+print
+print '<html>'
+print '<head><title>Hello from Python</title></head>'
+print '<body>'
+print '<h2>Hello from Python</h2>'
+print '</body></html>'
