+++ client/html/HomePage.html
@@ -5,21 +5,16 @@
 		<meta name = "viewport" content = "width = device-width, initial-scale = 1">
 		<title> Infinite Deals </title>
 		<link rel="stylesheet" href = "../css/bootstrap.min.css">
-		<link rel="stylesheet" type="text/css" href="../CSS/HomeStyle.css">
+		<link rel="stylesheet" type="text/css" href="../css/HomeStyle.css">
 
 	</head>
 	<body>
-		<!-- Java script file section -->
-		<script src= "../JavaScript/googleCSE.js" type="text/javascript"></script>
-		<gcse:search></gcse:search>
-
 		<div id = "warningBar">
 			  <strong>Disclaimer:</strong> Website were mean to use as personal project.
 			   All resources are copyright by original owner.
 		</div>
-
-		<img src= "../Img/MainLogo.png" alt= "Logo" class="center">
-		<form class="searchBar" action="../WebCrawler/NewEggs.py" method="get">
+		<img src= "../img/MainLogo.png" alt= "Logo" class="center">
+		<form class="searchBar" action="../../cgi-bin/NewEggs.py" method="get">
 			<input type="text" name="search" placeholder="Search.." class="center">
 		</form>
 	</body>
