+++ client/img/firewatch-game.jpg
